added controller for { getAllPokemons, getPokemonById, getPokemonInfo } and imported in index.js
