+++ backend/db/dbConnection.js
@@ -1,1 +1,9 @@
-//change the name if necessary
+//
+const mongoose = require ("mongoose"); 
+
+
+const connectMongoDB = async() => {
+    await mongoose.connect(process.env.MONGO_URI )
+};
+
+module.exports = connectMongoDB;
