+++ backend/package.json
@@ -16,6 +16,8 @@
     "demon": "^0.0.1",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "express-mongo-sanitize": "^2.2.0",
+    "mongoose": "^8.4.1",
     "no": "^0.0.1",
     "nodemon": "^3.1.3"
   }
